Move interface to fix Jenkins build error
